fix xcframework setup and move it
