+++ plugins/src/main/kotlin/com/freeletics/gradle/setup/PublishSetup.kt
@@ -2,13 +2,16 @@ package com.freeletics.gradle.setup
 
 import com.freeletics.gradle.util.freeleticsExtension
 import com.freeletics.gradle.util.stringProperty
-import com.github.javaparser.printer.lexicalpreservation.LexicalPreservingPrinter.setup
 import com.gradle.scan.agent.serialization.scan.serializer.kryo.it
 import com.vanniktech.maven.publish.MavenPublishBaseExtension
 import org.gradle.api.Project
 import org.gradle.api.credentials.PasswordCredentials
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPom
+import org.gradle.api.publish.maven.MavenPublication
+import org.gradle.api.publish.maven.tasks.AbstractPublishToMaven
+import org.gradle.api.tasks.bundling.Zip
+import org.jetbrains.kotlin.konan.target.HostManager
 
 internal fun setupOssPublishing(target: Project) {
     target.plugins.apply("org.jetbrains.dokka")
@@ -88,6 +91,41 @@ internal fun MavenPom.configurePom(project: Project, includeLicense: Boolean) {
                 license.url.set("https://www.apache.org/licenses/LICENSE-2.0.txt")
                 license.distribution.set("repo")
             }
+        }
+    }
+}
+
+internal fun setupXcFrameworkPublishing(project: Project, frameworkName: String) {
+    val framework = "$frameworkName.xcframework"
+    val frameworkRoot = project.layout.buildDirectory.dir("XCFrameworks/release")
+    val assembleTask = "assemble${frameworkName}ReleaseXCFramework"
+
+    val frameworkZip = project.tasks.register("${assembleTask}Zip", Zip::class.java) {
+        it.dependsOn(assembleTask)
+        it.onlyIf { HostManager.hostIsMac }
+
+        it.from(frameworkRoot.map { root -> root.dir(framework) })
+        it.into(framework)
+        it.archiveBaseName.set(framework)
+        it.destinationDirectory.set(frameworkRoot)
+        it.isPreserveFileTimestamps = false
+        it.isReproducibleFileOrder = true
+    }
+
+    val publicationName = "${frameworkName}XcFramework"
+    project.extensions.configure(PublishingExtension::class.java) { publishing ->
+        publishing.publications.create(publicationName, MavenPublication::class.java) {
+            // the project.name will be replaced with the real artifact id by the publishing plugin
+            it.artifactId = "${project.name}-xcframework"
+            it.artifact(frameworkZip) { artifact ->
+                artifact.extension = "zip"
+            }
+        }
+    }
+
+    project.tasks.withType(AbstractPublishToMaven::class.java).configureEach {
+        if (it.name.contains(publicationName, ignoreCase = true)) {
+            it.onlyIf { HostManager.hostIsMac }
         }
     }
 }
